server/requirements.txt
flask
python-dotenv
psycopg2-binary
Flask-Migrate
Flask-SQLAlchemy